+++ src/drivers/Makefile.header_from_source
@@ -1,12 +1,12 @@
 MODULE:= rtmouse
 obj-m:= $(MODULE).o
-$(MODULE)-y:= $(MODULE)_main.o $(MODULE)_dev_fops.o
+$(MODULE)-y:= $(MODULE)_main.o $(MODULE)_dev_fops.o $(MODULE)_spi.o
 clean-files:= *.o *.ko *.mod.[co] *~
 
 LINUX_SRC_DIR:=/usr/src/linux
 VERBOSE:=0
 
-$(MODULE).ko: $(MODULE)_dev_fops.c $(MODULE)_main.c $(MODULE).h
+$(MODULE).ko: $(MODULE)_main.c $(MODULE)_dev_fops.c $(MODULE)_spi.c $(MODULE).h
 	make -C $(LINUX_SRC_DIR) M=$(shell pwd) V=$(VERBOSE) modules
 
 clean:
