+++ src/drivers/rtmouse.h
@@ -262,10 +262,13 @@ extern struct file_operations dev_fops[ID_DEV_SIZE];
 extern struct device *mcp320x_dev;
 #endif
 
-/* --- used in rtmouse_dev_fops.c --- */
 int rpi_gpio_function_set(int pin, uint32_t func);
 void rpi_gpio_set32(uint32_t mask, uint32_t val);
 void rpi_gpio_clear32(uint32_t mask, uint32_t val);
 void rpi_pwm_write32(uint32_t offset, uint32_t val);
+
+/* --- used in rtmouse_spi.c --- */
+int mcp3204_init(void);
+void mcp3204_exit(void);
 
 #endif // RTMOUSE_H
